pyproject.toml
[project]
name = "content_crew"
version = "0.1.0"
description = "content_crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[google-genai,tools]==1.9.3",
    "scrapfly-sdk>=0.8.24",
    "scrapingbee>=2.0.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.100.0",
    "apscheduler>=3.10.0",
    "email-validator>=2.0.0",
]

[project.scripts]
content_crew = "content_crew.main:run"
run_crew = "content_crew.main:run"
train = "content_crew.main:train"
replay = "content_crew.main:replay"
test = "content_crew.main:test"
run_with_trigger = "content_crew.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
